S7 class to contain an OWL environment.

OWL_env is an S7 class which is typically slotted into OWL objects to serve as a context, an environment, in which to keep track of the "non-SPARQL query code" parts of the OWL object. This includes custom triple generating functions (terms), but also the variables named in the query so far.

Usage

OWL_env(x)

Arguments

x

List object.

Value

an OWL_env object.

Examples

OWL_env
#> <pallas::OWL_env> class
#> @ parent     : <list>
#> @ constructor: function(x) {...}
#> @ validator  : <NULL>
#> @ properties :