S7 class to contain an RDF triple.

triple is an S7 class contain RDF triples that make up a SPARQL query. A triple expression describes a relationship in three parts, known as subject–predicate–object expression. In the data class, this is reflected by its three slots.

Usage

triple(subject = c("", ""), predicate = c("", ""), object = c("", ""))

Arguments

subject

Character vector. Length one or two. First position gives subject name, second (optional, default NULL) gives prefix.

predicate

Character scalar. Length one or two. First position gives predicate name, second (optional, default NULL) gives prefix.

object

Character scalar. Length one or two. First position gives object name, second (optional, default NULL) gives prefix.

Value

a triple object.

Slots

subject

Character vector. Length one or two. First position gives subject name, second (optional, default NULL) gives prefix.

predicate

Character scalar. Length one or two. First position gives predicate name, second (optional, default NULL) gives prefix.

object

Character scalar. Length one or two. First position gives object name, second (optional, default NULL) gives prefix.

Examples

triple(
    subject = c("sub_prefix", "subject"),
    predicate = c("pred_prefix", "predicate"),
    object = c("obj_prefix", "object")
)
#> pallas::triple character S7_object.
#> sub_prefix subject   pred_prefix predicate   obj_prefix object